Reorganize code in snapshots package to facilitate windows implementation
