--- src/go/rdctl/pkg/snapshot/copyFile_windows.go
@@ -1,10 +1,0 @@
-package snapshot
-
-import (
-	"errors"
-	"os"
-)
-
-func copyFile(dst, src string, copyOnWrite bool, fileMode os.FileMode) error {
-	return errors.New("not implemented")
-}
