--- src/go/rdctl/pkg/snapshot/utils.go
@@ -1,17 +1,0 @@
-package snapshot
-
-import (
-	"math/rand"
-)
-
-// Returns a string of length n that is comprised of random letters
-// and numbers. From:
-// https://stackoverflow.com/questions/22892120/how-to-generate-a-random-string-of-a-fixed-length-in-go
-func randomString(n int) string {
-	letters := "abcdefghijklmnopqrstuvwxyz0123456789"
-	b := make([]byte, n)
-	for i := range b {
-		b[i] = letters[rand.Intn(len(letters))]
-	}
-	return string(b)
-}
